requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiofiles==23.2.1
email-validator==2.1.0
starlette==0.27.0
itsdangerous==2.1.2
